URL rewrite at the agent web server to support multiple dashboard versions (#15247)
* new routing for web requests

* renamed and better error control

* add missing return statements

* do not serve files when no file extension is given

* restore the api of the functions; use internal web_client flags to keep state; support redirects to fix directories

* add window.location.hash to url redirect

* do not redirect when sending to specific dashboard version and there are data after the version

* uniform function to append slash to URL

* remove obsolete proxy https flag

+++ aclk/aclk_query.c
@@ -128,7 +128,7 @@ static int http_api_v2(struct aclk_query_thread *query_thr, aclk_query_t query) 
         ACLK_STATS_UNLOCK;
     }
 
-    w->response.code = web_client_api_request_with_node_selection(localhost, w, path);
+    w->response.code = (short)web_client_api_request_with_node_selection(localhost, w, path);
     web_client_timeout_checkpoint_response_ready(w, &t);
 
     if(buffer_strlen(w->response.data) > ACLK_MAX_WEB_RESPONSE_SIZE) {
